create_order/app/core/config.py
from dotenv import load_dotenv
import os

load_dotenv()

DATABASE_URL=os.getenv("DATABASE_URL")
SECRET_KEY=os.getenv("SECRET_KEY")
ALGORITHM=os.getenv("ALGORITHM") 
LOCAL_DATABASE_URL = os.getenv("LOCAL_DATABASE_URL")